refactor(operator): :bell: add operator installation function
