+++ controllers/pkg/helm/submariner_operator.go
@@ -1,223 +1,90 @@
 package helm
 
-type CoreDNSCustomConfig struct{}
+import (
+	"context"
 
-type Images struct {
-	Repository string `yaml:"repository"`
-	Tag        string `yaml:"tag"`
-}
+	helmclient "github.com/mittwald/go-helm-client"
+	connectionhubv1alpha1 "github.com/robolaunch/connection-hub-operator/api/v1alpha1"
+	"k8s.io/client-go/rest"
+)
 
-type Submariner struct {
-	ClusterID           string              `yaml:"clusterId"`
-	Token               string              `yaml:"token"`
-	ClusterCIDR         string              `yaml:"clusterCidr"`
-	ServiceCIDR         string              `yaml:"serviceCidr"`
-	GlobalCIDR          string              `yaml:"globalCidr"`
-	NotEnabled          bool                `yaml:"natEnabled"`
-	ColorCodes          string              `yaml:"colorCodes"`
-	Debug               bool                `yaml:"debug"`
-	ServiceDiscovery    bool                `yaml:"serviceDiscovery"`
-	CableDriver         string              `yaml:"cableDriver"`
-	HealthCheckEnabled  bool                `yaml:"healthcheckEnabled"`
-	CoreDNSCustomConfig CoreDNSCustomConfig `yaml:"coreDNSCustomConfig"`
-	Images              Images              `yaml:"images"`
-}
+/*
+helm install submariner-operator  ./submariner-operator \
+--create-namespace --namespace "${SUBMARINER_NS}"  \
+--set submariner.clusterCidr="${CLUSTER_CIDR}" \
+--set submariner.serviceCidr="${SERVICE_CIDR}" \
+--set ipsec.psk="${SUBMARINER_PSK}" \
+--set broker.server="${SUBMARINER_BROKER_URL}" \
+--set broker.token="${SUBMARINER_BROKER_TOKEN}" \
+--set broker.namespace="${BROKER_NS}" \
+--set broker.ca="${SUBMARINER_BROKER_CA}" \
+--set submariner.serviceDiscovery=true \
+--set submariner.cableDriver=wireguard \
+--set submariner.clusterId="${CLUSTER_ID}" \
+--set submariner.natEnabled="true" \
+--set serviceAccounts.lighthouseAgent.create=true \
+--set serviceAccounts.lighthouseCoreDns.create=true \
+--set submariner.healthcheckEnabled=false \
+--set ipsec.natPort=4500 \
+--set ipsec.ikePort=500 \
+--set ipsec.preferredServer="true" \
+--set ipsec.natDiscovery=4490 \
+--set gateway.image.repository="docker.io/robolaunchio/submariner-gateway" \
+--set gateway.image.tag="dev-v11" \
+--set operator.image.repository="docker.io/robolaunchio/submariner-operator" \
+--set operator.image.tag="dev-v14" \
+--set submariner.images.repository="docker.io/robolaunchio" \
+--set submariner.images.tag="dev-v11"
+*/
 
-func getSubmarinerDefault() Submariner {
-	return Submariner{
-		ClusterID:           "",
-		Token:               "",
-		ClusterCIDR:         "",
-		ServiceCIDR:         "",
-		GlobalCIDR:          "",
-		NotEnabled:          false,
-		ColorCodes:          "blue",
-		Debug:               false,
-		ServiceDiscovery:    true,
-		CableDriver:         "libreswan",
-		HealthCheckEnabled:  true,
-		CoreDNSCustomConfig: CoreDNSCustomConfig{},
-		Images: Images{
-			Repository: "quay.io/submariner",
-			Tag:        "0.10.1",
-		},
+func InstallSubmarinerOperatorChart(submarinerOperator connectionhubv1alpha1.SubmarinerOperator, config *rest.Config) error {
+	cli, err := getClient(config, connectionhubv1alpha1.SubmarinerOperatorNamespace)
+	if err != nil {
+		return err
 	}
-}
 
-type Broker struct {
-	Server    string `yaml:"server"`
-	Token     string `yaml:"token"`
-	Namespace string `yaml:"namespace"`
-	Insecure  bool   `yaml:"insecure"`
-	Ca        string `yaml:"ca"`
-	GlobalNet bool   `yaml:"globalnet"`
-}
+	repoName := submarinerOperator.Spec.Helm.Repository.Name
+	repoURL := submarinerOperator.Spec.Helm.Repository.URL
 
-func getBrokerDefault() Broker {
-	return Broker{
-		Server:    "example.k8s.apiserver",
-		Token:     "test",
-		Namespace: "xyz",
-		Insecure:  false,
-		Ca:        "",
-		GlobalNet: false,
+	err = addRepository(config, connectionhubv1alpha1.SubmarinerOperatorNamespace, repoName, repoURL)
+	if err != nil {
+		return err
 	}
-}
 
-type RBAC struct {
-	Create bool `yaml:"create"`
-}
+	valuesObj := GetSubmarinerOperatorValuesDefault()
+	valuesObj.Submariner.ClusterCIDR = submarinerOperator.Spec.ClusterCIDR
+	valuesObj.Submariner.ServiceCIDR = submarinerOperator.Spec.ServiceCIDR
+	valuesObj.IPSEC.PSK = submarinerOperator.Spec.PresharedKey
+	valuesObj.Broker.Server = submarinerOperator.Spec.Broker.BrokerURL
+	valuesObj.Broker.Token = submarinerOperator.Spec.Broker.BrokerToken
+	valuesObj.Broker.Ca = submarinerOperator.Spec.Broker.BrokerCA
+	valuesObj.Submariner.ServiceDiscovery = true
+	valuesObj.Submariner.CableDriver = "wireguard"
+	valuesObj.Submariner.ClusterID = submarinerOperator.Spec.ClusterID
+	valuesObj.Submariner.NatEnabled = true
+	valuesObj.ServiceAccounts.LighthouseAgent.Create = true
+	valuesObj.ServiceAccounts.LighthouseCoreDNS.Create = true
+	valuesObj.Submariner.HealthCheckEnabled = false
+	valuesObj.IPSEC.NATPort = 4500
+	valuesObj.IPSEC.IKEPort = 500
+	valuesObj.IPSEC.PreferredServer = true
+	valuesObj.IPSEC.NATDiscovery = 4490
+	valuesObj.Gateway.Image.Repository = "docker.io/robolaunchio/submariner-gateway"
+	valuesObj.Gateway.Image.Tag = "dev-v11"
+	valuesObj.Operator.Image.Repository = "docker.io/robolaunchio/submariner-operator"
+	valuesObj.Operator.Image.Tag = "dev-v14"
+	valuesObj.Submariner.Images.Repository = "docker.io/robolaunchio"
+	valuesObj.Submariner.Images.Tag = "dev-v11"
 
-func getRBACDefault() RBAC {
-	return RBAC{
-		Create: true,
-	}
-}
-
-type IPSEC struct {
-	PSK             string `yaml:"psk"`
-	Debug           bool   `yaml:"debug"`
-	ForceUDPEncaps  bool   `yaml:"forceUDPEncaps"`
-	IKEPort         int    `yaml:"ikePort"`
-	NATPort         int    `yaml:"natPort"`
-	NATDiscovery    int    `yaml:"natDiscovery"`
-	PreferredServer bool   `yaml:"preferredServer"`
-}
-
-func getIPSECDefault() IPSEC {
-	return IPSEC{
-		PSK:             "",
-		Debug:           false,
-		ForceUDPEncaps:  false,
-		IKEPort:         500,
-		NATPort:         4500,
-		NATDiscovery:    4490,
-		PreferredServer: false,
-	}
-}
-
-type Leadership struct {
-	LeaseDuration int `yaml:"leaseDuration"`
-	RenewDeadline int `yaml:"renewDeadline"`
-	RetryPeriod   int `yaml:"retryPeriod"`
-}
-
-func getLeadershipDefault() Leadership {
-	return Leadership{
-		LeaseDuration: 10,
-		RenewDeadline: 5,
-		RetryPeriod:   2,
-	}
-}
-
-type OperatorImage struct {
-	Repository string `yaml:"repository"`
-	Tag        string `yaml:"tag"`
-	PullPolicy string `yaml:"pullPolicy"`
-}
-
-type OperatorResources struct{}
-
-type OperatorToleration struct{}
-
-type OperatorAffinity struct{}
-
-type Operator struct {
-	Image       OperatorImage        `yaml:"image"`
-	Resources   OperatorResources    `yaml:"resources"`
-	Tolerations []OperatorToleration `yaml:"tolerations"`
-	Affinity    OperatorAffinity     `yaml:"affinity"`
-}
-
-func getOperatorDefault() Operator {
-	return Operator{
-		Image: OperatorImage{
-			Repository: "quay.io/submariner/submariner-operator",
-			Tag:        "0.10.1",
-			PullPolicy: "IfNotPresent",
+	_, err = cli.InstallChart(
+		context.Background(),
+		&helmclient.ChartSpec{
+			ReleaseName: submarinerOperator.Spec.Helm.ReleaseName,
+			ChartName:   submarinerOperator.Spec.Helm.ChartName,
+			Version:     submarinerOperator.Spec.Helm.Version,
 		},
-		Resources:   OperatorResources{},
-		Tolerations: []OperatorToleration{},
-		Affinity:    OperatorAffinity{},
-	}
-}
+		&helmclient.GenericHelmOptions{},
+	)
 
-type Gateway struct {
-	Image Images `yaml:"image"`
-}
-
-func getOperatorGatewayDefault() Gateway {
-	return Gateway{
-		Image: Images{
-			Repository: "quay.io/submariner/submariner-gateway",
-			Tag:        "0.10.1",
-		},
-	}
-}
-
-type ServiceAccount struct {
-	Create bool   `yaml:"create"`
-	Name   string `yaml:"name"`
-}
-
-type ServiceAccounts struct {
-	Operator          ServiceAccount `yaml:"operator"`
-	Gateway           ServiceAccount `yaml:"gateway"`
-	RouteAgent        ServiceAccount `yaml:"routeAgent"`
-	GlobalNet         ServiceAccount `yaml:"globalnet"`
-	LighthouseAgent   ServiceAccount `yaml:"lighthouseAgent"`
-	LighthouseCoreDNS ServiceAccount `yaml:"lighthouseCoreDns"`
-}
-
-func getServiceAccountsDefault() ServiceAccounts {
-	return ServiceAccounts{
-		Operator: ServiceAccount{
-			Create: true,
-			Name:   "",
-		},
-		Gateway: ServiceAccount{
-			Create: true,
-			Name:   "",
-		},
-		RouteAgent: ServiceAccount{
-			Create: true,
-			Name:   "",
-		},
-		GlobalNet: ServiceAccount{
-			Create: true,
-			Name:   "",
-		},
-		LighthouseAgent: ServiceAccount{
-			Create: true,
-			Name:   "",
-		},
-		LighthouseCoreDNS: ServiceAccount{
-			Create: true,
-			Name:   "",
-		},
-	}
-}
-
-type SubmarinerOperatorValues struct {
-	Submariner      Submariner      `yaml:"submariner"`
-	Broker          Broker          `yaml:"broker"`
-	RBAC            RBAC            `yaml:"rbac"`
-	IPSEC           IPSEC           `yaml:"ipsec"`
-	Leadership      Leadership      `yaml:"leadership"`
-	Operator        Operator        `yaml:"operator"`
-	Gateway         Gateway         `yaml:"gateway"`
-	ServiceAccounts ServiceAccounts `yaml:"serviceAccounts"`
-}
-
-func GetSubmarinerOperatorValuesDefault() SubmarinerOperatorValues {
-	return SubmarinerOperatorValues{
-		Submariner:      getSubmarinerDefault(),
-		Broker:          getBrokerDefault(),
-		RBAC:            getRBACDefault(),
-		IPSEC:           getIPSECDefault(),
-		Leadership:      getLeadershipDefault(),
-		Operator:        getOperatorDefault(),
-		Gateway:         getOperatorGatewayDefault(),
-		ServiceAccounts: getServiceAccountsDefault(),
-	}
+	return err
 }
